backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
gspread
google-auth
google-api-python-client
python-multipart
email-validator
